requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --resolver=backtracking requirements/base.in
#
attrs==23.1.0
    # via
    #   cattrs
    #   jsii
aws-cdk-asset-awscli-v1==2.2.185
    # via aws-cdk-lib
aws-cdk-asset-kubectl-v20==2.1.1
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v5==2.0.155
    # via aws-cdk-lib
aws-cdk-lib==2.81.0
    # via -r requirements/base.in
aws-lambda-logging==0.1.1
    # via -r requirements/../function_requirements/common-requirements.txt
awscli==1.27.142
    # via -r requirements/base.in
boto3==1.26.142
    # via
    #   -r requirements/../function_requirements/common-requirements.txt
    #   -r requirements/base.in
botocore==1.29.142
    # via
    #   -r requirements/../function_requirements/common-requirements.txt
    #   -r requirements/base.in
    #   awscli
    #   boto3
    #   s3transfer
cachetools==5.3.0
    # via
    #   -r requirements/../function_requirements/common-requirements.txt
    #   google-auth
cattrs==22.2.0
    # via jsii
certifi==2023.5.7
    # via
    #   -r requirements/../function_requirements/common-requirements.txt
    #   requests
charset-normalizer==3.1.0
    # via
    #   -r requirements/../function_requirements/common-requirements.txt
    #   requests
colorama==0.4.4
    # via awscli
constructs==10.2.37
    # via
    #   -r requirements/base.in
    #   aws-cdk-lib
docutils==0.16
    # via awscli
exceptiongroup==1.1.1
    # via cattrs
google-api-core==2.11.0
    # via
    #   -r requirements/../function_requirements/common-requirements.txt
    #   google-api-python-client
google-api-python-client==2.87.0
    # via -r requirements/../function_requirements/common-requirements.txt
google-auth==2.19.0
    # via
    #   -r requirements/../function_requirements/common-requirements.txt
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.1.0
    # via
    #   -r requirements/../function_requirements/common-requirements.txt
    #   google-api-python-client
google-auth-oauthlib==1.0.0
    # via -r requirements/../function_requirements/common-requirements.txt
googleapis-common-protos==1.59.0
    # via
    #   -r requirements/../function_requirements/common-requirements.txt
    #   google-api-core
httplib2==0.22.0
    # via
    #   -r requirements/../function_requirements/common-requirements.txt
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.4
    # via
    #   -r requirements/../function_requirements/common-requirements.txt
    #   requests
importlib-resources==5.12.0
    # via jsii
invoke==2.1.2
    # via -r requirements/base.in
jmespath==1.0.1
    # via
    #   -r requirements/../function_requirements/common-requirements.txt
    #   boto3
    #   botocore
jsii==1.82.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-lib
    #   constructs
oauthlib==3.2.2
    # via
    #   -r requirements/../function_requirements/common-requirements.txt
    #   requests-oauthlib
protobuf==4.23.2
    # via
    #   -r requirements/../function_requirements/common-requirements.txt
    #   google-api-core
    #   googleapis-common-protos
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-lib
    #   constructs
    #   jsii
pyasn1==0.5.0
    # via
    #   -r requirements/../function_requirements/common-requirements.txt
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via
    #   -r requirements/../function_requirements/common-requirements.txt
    #   google-auth
pyjwt==2.7.0
    # via -r requirements/../function_requirements/common-requirements.txt
pyparsing==3.0.9
    # via
    #   -r requirements/../function_requirements/common-requirements.txt
    #   httplib2
python-dateutil==2.8.2
    # via
    #   -r requirements/../function_requirements/common-requirements.txt
    #   botocore
    #   jsii
python-dotenv==1.0.0
    # via
    #   -r requirements/../function_requirements/common-requirements.txt
    #   -r requirements/base.in
pytz==2023.3
    # via -r requirements/../function_requirements/common-requirements.txt
pyyaml==5.4.1
    # via awscli
requests==2.31.0
    # via
    #   -r requirements/../function_requirements/common-requirements.txt
    #   -r requirements/base.in
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via
    #   -r requirements/../function_requirements/common-requirements.txt
    #   google-auth-oauthlib
rsa==4.7.2
    # via
    #   -r requirements/../function_requirements/common-requirements.txt
    #   awscli
    #   google-auth
s3transfer==0.6.1
    # via
    #   -r requirements/../function_requirements/common-requirements.txt
    #   awscli
    #   boto3
six==1.16.0
    # via
    #   -r requirements/../function_requirements/common-requirements.txt
    #   google-auth
    #   google-auth-httplib2
    #   python-dateutil
tabulate==0.9.0
    # via -r requirements/base.in
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-lib
    #   constructs
    #   jsii
typing-extensions==4.6.2
    # via jsii
uritemplate==4.1.1
    # via
    #   -r requirements/../function_requirements/common-requirements.txt
    #   google-api-python-client
urllib3==1.26.16
    # via
    #   -r requirements/../function_requirements/common-requirements.txt
    #   botocore
    #   google-auth
    #   requests
zipp==3.15.0
    # via importlib-resources
